stock_research/StockAnalysis/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
tenacity==9.1.4
python-dotenv==1.2.2
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
fastapi==0.135.1
uvicorn[standard]==0.41.0
tradingview-scraper>=0.4.20
tradingview-screener>=2.5.0
